Added an initializer that sets the specific delegates up.
The actual setup process consists of several steps:
- See if the config file exists, and contains a dictionary.
- Take that dictionary, and validate its contents.
- Take the sanitized dictionary, and apply those settings, while (again) sanitizing what’s being set up.

This commit handles the last two steps, and populates the specific delegates based on the dictionary.
In order to have this testable, these steps happen in a separate initializer.
Since the class is currently implemented as a singleton, this initializer is not pubic, though!

Because the main reason for a singleton was ease of access to the One True Delegate© for the purpose of setting all the specific delegates, this implementation detail can now be removed in a later commit, and this new initializer can be published as the designated one.

The next commit will handle the remaining initial step of the setup, in the implementation of -init.