Added a dictionary for convenient method forwarding, and helper functions to manage it.
Objective-C has a very powerful forwarding mechanism that we can leverage to dispatch the message to their respective targets.
All we need to know is when we may forward which message to which of our specific delegate properties.
This can trivially be implemented using a CFDictionary where each key is a selector, and the values are the object that actually implements that selector.
Because selectors are uniqued by the runtime, the keys can be compared using pointer equality.
This spares us from implementing any key callbacks \o/

The next commit will delete all the old helper functions, and replace the implementations of the delegate methods by message forwarding.